+++ components/archive/shared/src/main/java/com/flipperdevices/archive/shared/composable/ComposableKeyCard.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import com.flipperdevices.archive.shared.utils.ExtractKeyMetaInformation
 import com.flipperdevices.bridge.dao.api.model.FlipperKeyType.Companion.colorByFlipperKeyType
 import com.flipperdevices.bridge.dao.api.model.parsed.FlipperKeyParsed
-import com.flipperdevices.core.ui.ktx.ComposableKeyType
 import com.flipperdevices.core.ui.ktx.clickableRipple
+import com.flipperdevices.core.ui.ktx.elements.ComposableKeyType
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography