+++ components/bridge/synchronization/ui/src/main/java/com/flipperdevices/bridge/synchronization/ui/composable/ComposableItemSynchronizationState.kt
@@ -16,8 +16,8 @@ import androidx.compose.ui.draw.rotate
 import androidx.compose.ui.res.stringResource
 import com.flipperdevices.bridge.synchronization.ui.R
 import com.flipperdevices.bridge.synchronization.ui.model.ItemSynchronizationState
-import com.flipperdevices.core.ui.ktx.animatedDots
-import com.flipperdevices.core.ui.ktx.painterResourceByKey
+import com.flipperdevices.core.ui.ktx.elements.animatedDots
+import com.flipperdevices.core.ui.ktx.image.painterResourceByKey
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography
 import com.flipperdevices.core.ui.res.R as DesignSystem