+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/api/FapVersionApiImpl.kt
@@ -5,18 +5,18 @@ import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.faphub.dao.api.FapVersionApi
 import com.flipperdevices.faphub.dao.api.model.FapItemVersion
-import com.flipperdevices.faphub.dao.network.retrofit.api.RetrofitVersionApi
+import com.flipperdevices.faphub.dao.network.retrofit.api.KtorfitVersionApi
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, FapVersionApi::class)
 class FapVersionApiImpl @Inject constructor(
-    private val retrofitVersionApi: RetrofitVersionApi
+    private val ktorfitVersionApi: KtorfitVersionApi
 ) : FapVersionApi, LogTagProvider {
     override val TAG = "FapVersionApi"
 
     override suspend fun getVersions(versions: List<String>): List<FapItemVersion> {
-        return retrofitVersionApi.getVersions(versions).map {
+        return ktorfitVersionApi.getVersions(versions).map {
             FapItemVersion(
                 id = it.id,
                 version = SemVer.fromString(it.version) ?: error("Failed parse ${it.version}")