+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/retrofit/api/KtorfitVersionApi.kt
@@ -1,0 +1,12 @@
+package com.flipperdevices.faphub.dao.network.retrofit.api
+
+import com.flipperdevices.faphub.dao.network.retrofit.model.RetrofitDetailedVersion
+import de.jensklingenberg.ktorfit.http.GET
+import de.jensklingenberg.ktorfit.http.Query
+
+interface KtorfitVersionApi {
+    @GET("application/versions")
+    suspend fun getVersions(
+        @Query("uid") ids: List<String>
+    ): List<RetrofitDetailedVersion>
+}