+++ components/keyemulate/api/src/main/kotlin/com/flipperdevices/keyemulate/api/KeyEmulateApi.kt
@@ -1,0 +1,10 @@
+package com.flipperdevices.keyemulate.api
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.flipperdevices.bridge.dao.api.model.FlipperKey
+
+interface KeyEmulateApi {
+    @Composable
+    fun ComposableEmulateButton(modifier: Modifier, flipperKey: FlipperKey)
+}