+++ components/updater/card/src/main/java/com/flipperdevices/updater/card/composable/ComposableUpdaterReboot.kt
@@ -13,7 +13,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.flipperdevices.core.ui.ktx.animatedDots
+import com.flipperdevices.core.ui.ktx.elements.animatedDots
 import com.flipperdevices.core.ui.theme.FlipperThemeInternal
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.core.ui.theme.LocalTypography