+++ components/widget/impl/src/main/java/com/flipperdevices/widget/impl/tasks/StartEmulateWorker.kt
@@ -15,8 +15,8 @@ import com.flipperdevices.core.ktx.jre.withCoroutineScope
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.error
 import com.flipperdevices.core.log.info
-import com.flipperdevices.keyscreen.api.emulate.AlreadyOpenedAppException
-import com.flipperdevices.keyscreen.api.emulate.EmulateHelper
+import com.flipperdevices.keyemulate.api.EmulateHelper
+import com.flipperdevices.keyemulate.exception.AlreadyOpenedAppException
 import com.flipperdevices.widget.impl.di.WidgetComponent
 import com.flipperdevices.widget.impl.model.WidgetState
 import com.flipperdevices.widget.impl.storage.WidgetStateStorage